+++ ca-mgmt-client/src/main/java/org/xipki/ca/mgmt/db/port/CaCertstore.java
@@ -372,12 +372,22 @@ public class CaCertstore extends ValidatableConf {
 
     private int profileId;
 
+    private String aliases;
+
     public int getProfileId() {
       return profileId;
     }
 
     public void setProfileId(int profileId) {
       this.profileId = profileId;
+    }
+
+    public String getAliases() {
+      return aliases;
+    }
+
+    public void setAliases(String aliases) {
+      this.aliases = aliases;
     }
 
     @Override