+++ CHANGELOG.md
@@ -10,16 +10,17 @@ See also <https://github.com/xipki/xipki/releases>
   - Feature: add limitation to the name of CA, publisher, requestor, cert profile, signer, and alias of CA.
   - Feature: add support of constant value of types PrintableString, UTF8String, INTEGER, BIT STRING and OCTET STRING.
   - Feature: add limitation to the name of CAs, signers, publishers, requestors, and certificate profiles.
-  - Add support of tomcat 10
+  - Feature: allow the use of aliases for certificate profiles in a CA.
+  - Add support of tomcat 10+
 - OCSP
   - Feature: extend Properties to use the place-holder ${env:name} for environment and ${sys:name} for system property.
-  - Add support of tomcat 10
+  - Add support of tomcat 10+
 - Gateway
   - Feature: add support of ACME with challenge types dns-01, http-01 and tls-alpn-01
   - Feature: encode the requests and responses between gateway and ca-server in CBOR format (was JSON).
   - Feature: extend Properties to use the place-holder ${env:name} for environment and ${sys:name} for system property.
   - Feature: add support of short URLs in EST, REST and SCEP gateways.
-  - Add support of tomcat 10
+  - Add support of tomcat 10+
 - CLI
   - N/A
 - MGMT-CLI (Management Client)